refactor: Refactor authentication lib
