+++ packages/webapi/models/mock/AuthInfoModel.json
@@ -1,0 +1,4 @@
+{
+	"scheme": "jwt",
+	"authURL": "/auth"
+}
