+++ packages/webapi/models/mock/LoginRequest.json
@@ -1,0 +1,4 @@
+{
+	"username":"wasp",
+	"password":"wasp"
+}
